+++ pysnmp/proto/rfc1902.py
@@ -177,7 +177,7 @@ class Integer(Integer32):
         enums.update(values.items())
 
         class X(cls):
-            named_values = namedval.NamedValues(*enums)
+            namedValues = namedval.NamedValues(*enums)
             subtypeSpec = cls.subtypeSpec + constraint.SingleValueConstraint(
                 *values.values()
             )  # noqa: N815
@@ -234,20 +234,20 @@ class OctetString(univ.OctetString):
     # having zero-range size constraint applied. The following is
     # supposed to be used for setting and querying this property.
 
-    fixed_length = None
+    fixedLength = None
 
     def set_fixed_length(self, value):
         """Set fixed length."""
-        self.fixed_length = value
+        self.fixedLength = value
         return self
 
     def is_fixed_length(self):
         """Return if fixed length."""
-        return self.fixed_length is not None
+        return self.fixedLength is not None
 
     def get_fixed_length(self):
         """Return fixed length."""
-        return self.fixed_length
+        return self.fixedLength
 
     def clone(self, *args, **kwargs):
         """Clone the data."""
@@ -367,7 +367,7 @@ class IpAddress(OctetString):
     subtypeSpec = OctetString.subtypeSpec + constraint.ValueSizeConstraint(
         4, 4
     )  # noqa: N815
-    fixed_length = 4
+    fixedLength = 4
 
     def prettyIn(self, value):  # noqa: N802
         """Convert string to IP address."""
@@ -683,7 +683,7 @@ class Bits(OctetString):
 
     """
 
-    named_values = namedval.NamedValues()
+    namedValues: namedval.NamedValues = namedval.NamedValues()
 
     def __new__(cls, *args, **kwargs):
         """Create a new instance of the class."""
@@ -699,7 +699,7 @@ class Bits(OctetString):
             return OctetString.prettyIn(self, bits)  # raw bitstring
         octets = []
         for bit in bits:  # tuple of named bits
-            v = self.named_values.getValue(bit)
+            v = self.namedValues.getValue(bit)
             if v is None:
                 raise error.ProtocolError("Unknown named bit %s" % bit)
             d, m = divmod(v, 8)
@@ -717,7 +717,7 @@ class Bits(OctetString):
             j = 7
             while j >= 0:
                 if v & (0x01 << j):
-                    name = self.named_values.getName(i * 8 + 7 - j)
+                    name = self.namedValues.getName(i * 8 + 7 - j)
                     if name is None:
                         name = f"UnknownBit-{i * 8 + 7 - j}"
                     names.append(name)
@@ -730,11 +730,11 @@ class Bits(OctetString):
 
         Reduce fully duplicate enumerations along the way.
         """
-        enums = set(cls.named_values.items())
+        enums = set(cls.namedValues.items())
         enums.update(values.items())
 
         class X(cls):
-            named_values = namedval.NamedValues(*enums)
+            namedValues = namedval.NamedValues(*enums)
 
         X.__name__ = cls.__name__
         return X