+++ tests/smi/manager/test_mib-tree-inspection.py
@@ -1,10 +1,17 @@
 import pytest
-from pysnmp.smi import builder, view
+from pysnmp.smi import builder, compiler, view
 
 mibBuilder = builder.MibBuilder()
-mibBuilder.add_mib_sources(builder.DirMibSource("/opt/pysnmp_mibs"))
-mibBuilder.load_modules("SNMPv2-MIB", "SNMP-FRAMEWORK-MIB", "SNMP-COMMUNITY-MIB")
+mibBuilder.loadTexts = True
+compiler.addMibCompiler(mibBuilder)
 mibView = view.MibViewController(mibBuilder)
+
+mibBuilder.load_modules(
+    "SNMPv2-MIB",
+    "SNMP-FRAMEWORK-MIB",
+    "SNMP-COMMUNITY-MIB",
+    "CISCO-ENHANCED-IPSEC-FLOW-MIB",
+)
 
 
 def test_getNodeName_by_OID():
@@ -73,7 +80,7 @@ def test_getNodeName_by_symbol_description_with_module_name():
     assert suffix == ()
 
     (mibNode,) = mibBuilder.import_symbols("SNMP-FRAMEWORK-MIB", "snmpEngineID")
-    assert mibNode.syntax.prettyPrint() != ""
+    assert "" != mibNode.syntax.prettyPrint()
 
 
 def test_getNodeName_by_symbol_location_lookup_by_name():
@@ -81,3 +88,34 @@ def test_getNodeName_by_symbol_location_lookup_by_name():
     assert modName == "SNMP-COMMUNITY-MIB"
     assert symName == "snmpCommunityEntry"
     assert suffix == ()
+
+
+def test_getNodeName_by_symbol_description_with_module_name_2():
+    oid, label, suffix = mibView.get_node_name(
+        ("ciscoEnhIPsecFlowActivityGroup",), "CISCO-ENHANCED-IPSEC-FLOW-MIB"
+    )
+    assert oid == (1, 3, 6, 1, 4, 1, 9, 9, 432, 2, 2, 1)
+    assert label == (
+        "iso",
+        "org",
+        "dod",
+        "internet",
+        "private",
+        "enterprises",
+        "cisco",
+        "ciscoMgmt",
+        "ciscoEnhancedIpsecFlowMIB",
+        "ciscoEnhancedIpsecFlowMIBConform",
+        "ciscoIPsecFlowMIBGroups",
+        "ciscoEnhIPsecFlowActivityGroup",
+    )
+    assert suffix == ()
+
+    (mibNode,) = mibBuilder.import_symbols(
+        "CISCO-ENHANCED-IPSEC-FLOW-MIB", "ciscoEnhIPsecFlowActivityGroup"
+    )
+    assert (
+        "This group consists of: 1) IPsec Phase-2 Global Statistics 2) IPsec Phase-2 Tunnel Table 3) IPsec Phase-2 Endpoint Table 4) IPsec Phase-2 Security Association Table"
+        == mibNode.getDescription()
+    )
+    assert "rfc2408, rfc2407; rfc2409 section 5.5" == mibNode.getReference()